--- examples/21-singularity/21-singularity.cpp
@@ -1,257 +1,0 @@
-/*
- * Example of singularity handling by smoothing Lambda in/out of singularities.
- */
-
-#include <math.h>
-#include <signal.h>
-
-#include <iostream>
-#include <mutex>
-#include <string>
-#include <thread>
-
-#include "Sai2Graphics.h"
-#include "Sai2Model.h"
-#include "Sai2Simulation.h"
-#include "tasks/JointTask.h"
-#include "tasks/MotionForceTask.h"
-#include "timer/LoopTimer.h"
-#include "logger/Logger.h"
-bool fSimulationRunning = false;
-void sighandler(int) { fSimulationRunning = false; }
-
-using namespace std;
-using namespace Eigen;
-
-const string world_file = "resources/world.urdf";
-const string robot_file = "resources/panda_arm.urdf";
-const string robot_name = "PANDA";
-
-// ui torques and control torques
-VectorXd ui_torques;
-VectorXd control_torques;
-
-// mutex for global variables between different threads
-mutex mutex_torques;
-mutex mutex_robot;
-
-// simulation and control loop
-void control(shared_ptr<Sai2Model::Sai2Model> robot,
-			 shared_ptr<Sai2Simulation::Sai2Simulation> sim);
-void simulation(shared_ptr<Sai2Model::Sai2Model> robot,
-				shared_ptr<Sai2Simulation::Sai2Simulation> sim);
-
-//------------ main function
-int main(int argc, char** argv) {
-	cout << "Loading URDF world model file: " << world_file << endl;
-
-	// set up signal handler
-	signal(SIGABRT, &sighandler);
-	signal(SIGTERM, &sighandler);
-	signal(SIGINT, &sighandler);
-
-	// load graphics scene
-	auto graphics = make_shared<Sai2Graphics::Sai2Graphics>(world_file);
-	graphics->addUIForceInteraction(robot_name);
-	// graphics->showTransparency(true, robot_name, 0.5);
-
-	// load simulation world
-	auto sim = make_shared<Sai2Simulation::Sai2Simulation>(world_file);
-
-	// load robots
-	auto robot = make_shared<Sai2Model::Sai2Model>(robot_file, false);
-	robot->setQ(sim->getJointPositions(robot_name));
-	robot->updateModel();
-
-	// sim->setJointPositions(robot_name, 0 * robot->q());
-
-	// intitialize global torques variables
-	ui_torques = VectorXd::Zero(robot->dof());
-	control_torques = VectorXd::Zero(robot->dof());
-
-	// start the simulation thread first
-	fSimulationRunning = true;
-	thread sim_thread(simulation, robot, sim);
-
-	// next start the control thread
-	thread ctrl_thread(control, robot, sim);
-
-	// while window is open:
-	while (graphics->isWindowOpen()) {
-		{
-			lock_guard<mutex> lock(mutex_robot);
-			graphics->updateRobotGraphics(robot_name, robot->q());
-		}
-		graphics->renderGraphicsWorld();
-		{
-			lock_guard<mutex> lock(mutex_torques);
-			ui_torques = graphics->getUITorques(robot_name);
-		}
-	}
-
-	// stop simulation
-	fSimulationRunning = false;
-	sim_thread.join();
-	ctrl_thread.join();
-
-	return 0;
-}
-
-//------------------------------------------------------------------------------
-void control(shared_ptr<Sai2Model::Sai2Model> robot,
-			 shared_ptr<Sai2Simulation::Sai2Simulation> sim) {
-	// update robot model and initialize control vectors
-	robot->updateModel();
-	int dof = robot->dof();
-	MatrixXd N_prec = MatrixXd::Identity(dof, dof);
-
-	// Position plus orientation task
-	string link_name = "end-effector";
-	Vector3d pos_in_link = Vector3d(0.0, 0.0, 0.07);
-	Affine3d compliant_frame = Affine3d(Translation3d(pos_in_link));
-
-	// Full motion force task
-	auto motion_force_task = make_unique<Sai2Primitives::MotionForceTask>(
-		robot, link_name, compliant_frame);
-
-	// // Partial motion force task
-	// vector<Vector3d> controlled_directions_translation = {
-	// 	Vector3d::UnitX(), Vector3d::UnitY(), Vector3d::UnitZ()};
-	// vector<Vector3d> controlled_directions_rotation = {};
-	// auto motion_force_task = make_shared<Sai2Primitives::MotionForceTask>(
-	// 	robot, link_name, controlled_directions_translation,
-	// 	controlled_directions_rotation);
-	// motion_force_task->setSingularityBounds(5e-2, 5e-3);
-	// motion_force_task->setSingularityGains(20, 20);
-
-    motion_force_task->disableInternalOtg();
-    motion_force_task->enableVelocitySaturation();
-	VectorXd motion_force_task_torques = VectorXd::Zero(dof);
-
-	// no gains setting here, using the default task values
-	const Matrix3d initial_orientation = robot->rotation(link_name);
-	const Vector3d initial_position = robot->position(link_name, pos_in_link);
-
-	// joint task to control the redundancy
-	// using default gains and interpolation settings
-	auto joint_task = make_unique<Sai2Primitives::JointTask>(robot);
-    joint_task->setGains(100, 20);
-	VectorXd joint_task_torques = VectorXd::Zero(dof);
-
-	VectorXd initial_q = robot->q();
-    joint_task->setGoalPosition(initial_q);
-
-    // desired position offsets 
-    vector<Vector3d> desired_offsets {Vector3d(2, 0, 0), Vector3d(0, 0, 0), 
-                                      Vector3d(0, 2, 0), Vector3d(0, 0, 0), 
-                                      Vector3d(0, -2, 0), Vector3d(0, 0, 0),
-                                      Vector3d(0, 0, 2), Vector3d(0, 0, 0)};
-    // vector<Vector3d> desired_offsets {Vector3d(2, 0, 0)};
-	double t_initial = 2;
-	vector<double> t_wait {5, 5};
-    // double t_wait = 10;  // wait between switching desired positions 
-	// double t_reset_wait = 5;  // wait when resetting position 
-    double prev_time = 0;
-    // int cnt = 6 * 1;
-	int cnt = 0;
-    int max_cnt = desired_offsets.size();
-
-	// create logger
-	Sai2Common::Logger logger("singular_values", false);
-	VectorXd svalues = VectorXd::Zero(6);
-	logger.addToLog(svalues, "svalues");
-	logger.start();
-
-	// create a loop timer
-	double control_freq = 1000;
-	Sai2Common::LoopTimer timer(control_freq, 1e6);
-
-	while (fSimulationRunning) {
-		timer.waitForNextLoop();
-		const double time = timer.elapsedSimTime();
-
-		// read joint positions, velocities, update model
-		robot->setQ(sim->getJointPositions(robot_name));
-		robot->setDq(sim->getJointVelocities(robot_name));
-		robot->updateModel();
-
-		// update tasks model. Order is important to define the hierarchy
-		N_prec = MatrixXd::Identity(dof, dof);
-		{
-			lock_guard<mutex> lock(mutex_robot);
-			motion_force_task->updateTaskModel(N_prec);
-		}
-		N_prec = motion_force_task->getTaskAndPreviousNullspace();
-		// after each task, need to update the nullspace
-		// of the previous tasks in order to garantee
-		// the dyamic consistency
-
-		joint_task->updateTaskModel(N_prec);
-
-		// -------- set task goals and compute control torques
-		// position: move to workspace extents 
-        if (time - prev_time > t_wait[cnt % 2]) {
-            motion_force_task->setGoalPosition(initial_position + desired_offsets[cnt]);
-            cnt++;
-            prev_time = time;
-            if (cnt == max_cnt) cnt = max_cnt - 1;
-        }
-        motion_force_task->setGoalLinearVelocity(Vector3d::Zero());
-        motion_force_task->setGoalLinearAcceleration(Vector3d::Zero());
-
-		// compute torques for the different tasks
-		motion_force_task_torques = motion_force_task->computeTorques();
-		joint_task_torques = joint_task->computeTorques();
-
-		//------ compute the final torques
-		{
-			lock_guard<mutex> lock(mutex_torques);
-			control_torques = motion_force_task_torques + joint_task_torques;
-		}
-
-		//------ logging data
-		auto svd_data = motion_force_task->getSingularitySvdData();
-		svalues = svd_data.s;
-
-		// -------------------------------------------
-		if (timer.elapsedCycles() % 500 == 0) {
-			cout << "time: " << time << endl;
-			cout << "position error : "
-				 << (motion_force_task->getGoalPosition() -
-					 motion_force_task->getCurrentPosition())
-						.norm()
-				 << endl;
-			cout << endl;
-		}
-	}
-	logger.stop();
-	timer.stop();
-	cout << "\nControl loop timer stats:\n";
-	timer.printInfoPostRun();
-}
-
-//------------------------------------------------------------------------------
-void simulation(shared_ptr<Sai2Model::Sai2Model> robot,
-				shared_ptr<Sai2Simulation::Sai2Simulation> sim) {
-	fSimulationRunning = true;
-
-	sim->disableJointLimits(robot_name);
-
-	// create a timer
-	double sim_freq = 2000;
-	Sai2Common::LoopTimer timer(sim_freq);
-
-	sim->setTimestep(1.0 / sim_freq);
-
-	while (fSimulationRunning) {
-		timer.waitForNextLoop();
-		{
-			lock_guard<mutex> lock(mutex_torques);
-			sim->setJointTorques(robot_name, control_torques + ui_torques);
-		}
-		sim->integrate();
-	}
-	timer.stop();
-	cout << "\nSimulation loop timer stats:\n";
-	timer.printInfoPostRun();
-}
