--- examples/21-singularity/panda_arm.urdf
@@ -1,185 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<robot name="panda">
-
-    <link name="link0">
-      <inertial>
-        <origin xyz="0 0 0.05" rpy="0 0 0" />
-        <mass value="4" />
-        <inertia ixx="0.4" iyy="0.4" izz="0.4" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link0.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link1">
-      <inertial>
-        <origin xyz="0 0 -0.07" rpy="0 -0 0" />
-        <mass value="3" />
-        <inertia ixx="0.3" iyy="0.3" izz="0.3" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link1.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link2">
-      <inertial>
-        <origin xyz="0 -0.1 0" rpy="0 -0 0" />
-        <mass value="3" />
-        <inertia ixx="0.3" iyy="0.3" izz="0.3" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link2.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link3">
-      <inertial>
-        <origin xyz="0.04 0 -0.05" rpy="0 -0 0" />
-        <mass value="2" />
-        <inertia ixx="0.2" iyy="0.2" izz="0.2" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link3.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link4">
-      <inertial>
-        <origin xyz="-0.04 0.05 0" rpy="0 -0 0" />
-        <mass value="2" />
-        <inertia ixx="0.2" iyy="0.2" izz="0.2" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link4.obj"/>
-        </geometry>
-      </visual>   
-    </link>
-    <link name="link5">
-      <inertial>
-        <origin xyz="0 0 -0.15" rpy="0 -0 0" />
-        <mass value="2" />
-        <inertia ixx="0.2" iyy="0.2" izz="0.2" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link5.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link6">
-      <inertial>
-        <origin xyz="0.06 0 0" rpy="0 -0 0" />
-        <mass value="1.5" />
-        <inertia ixx="0.1" iyy="0.1" izz="0.1" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link6.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="link7">
-      <inertial>
-        <origin xyz="0 0 0.17" rpy="0 -0 0" />
-        <mass value="1.8" />
-        <inertia ixx="0.09" iyy="0.05" izz="0.07" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <mesh filename="../../../../examples/models/panda/meshes/visual/link7.obj"/>
-        </geometry>
-      </visual>
-    </link>
-    <link name="end-effector">
-      <inertial>
-        <origin xyz="0 0 0" rpy="0 -0 0" />
-        <mass value="0.2" />
-        <inertia ixx="0.01" iyy="0.01" izz="0.01" ixy="0" ixz="0" iyz="0" />
-      </inertial>
-      <visual>
-        <geometry>
-          <sphere radius="0.04" />
-        </geometry>
-      </visual>
-      <collision>
-        <geometry>
-          <sphere radius="0.04" />
-        </geometry>
-      </collision>
-    </link>
-  
-    <joint name="joint1" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-2.8973" soft_upper_limit="2.8973"/>
-      <origin rpy="0 0 0" xyz="0 0 0.333"/>
-      <parent link="link0"/>
-      <child link="link1"/>
-      <axis xyz="0 0 1"/>
-      <calibration falling="0.0" />
-      <limit effort="87" lower="-2.8973" upper="2.8973" velocity="2.1750"/>
-    </joint>
-    <joint name="joint2" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-1.7628" soft_upper_limit="1.7628"/>
-      <origin rpy="-1.57079632679 0 0" xyz="0 0 0"/>
-      <parent link="link1"/>
-      <child link="link2"/>
-      <axis xyz="0 0 1"/>
-      <calibration falling="-25.0" />
-      <limit effort="87" lower="-1.7628" upper="1.7628" velocity="2.1750"/>
-    </joint>
-    <joint name="joint3" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-2.8973" soft_upper_limit="2.8973"/>
-      <origin rpy="1.57079632679 0 0" xyz="0 -0.316 0"/>
-      <parent link="link2"/>
-      <child link="link3"/>
-      <axis xyz="0 0 1"/>
-      <limit effort="87" lower="-2.8973" upper="2.8973" velocity="2.1750"/>
-    </joint>
-    <joint name="joint4" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-3.0718" soft_upper_limit="-0.0698"/>
-      <origin rpy="1.57079632679 0 0" xyz="0.0825 0 0"/>
-      <parent link="link3"/>
-      <child link="link4"/>
-      <axis xyz="0 0 1"/>
-      <calibration falling="-135.0" />
-      <!-- <limit effort="87" lower="-3.0718" upper="-0.0698" velocity="2.1750"/> -->
-      <limit effort="87" lower="-3.0718" upper="-0.5" velocity="2.1750"/>
-    </joint>
-    <joint name="joint5" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-2.8973" soft_upper_limit="2.8973"/>
-      <origin rpy="-1.57079632679 0 0" xyz="-0.0825 0.384 0"/>
-      <parent link="link4"/>
-      <child link="link5"/>
-      <axis xyz="0 0 1"/>
-      <calibration falling="0.0" />
-      <limit effort="12" lower="-2.8973" upper="2.8973" velocity="2.6100"/>
-    </joint>
-    <joint name="joint6" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-0.0175" soft_upper_limit="3.7525"/>
-      <origin rpy="1.57079632679 0 0" xyz="0 0 0"/>
-      <parent link="link5"/>
-      <child link="link6"/>
-      <axis xyz="0 0 1"/>
-      <calibration falling="105.0" />
-      <limit effort="12" lower="-0.0175" upper="3.7525" velocity="2.6100"/>
-    </joint>
-    <joint name="joint7" type="revolute">
-      <safety_controller k_position="100.0" k_velocity="40.0" soft_lower_limit="-2.8973" soft_upper_limit="2.8973"/>
-      <origin rpy="1.57079632679 0 0" xyz="0.088 0 0"/>
-      <parent link="link6"/>
-      <child link="link7"/>
-      <axis xyz="0 0 1"/>
-      <limit effort="12" lower="-2.8973" upper="2.8973" velocity="2.6100"/>
-    </joint>
-    <joint name="joint_ee" type="fixed">
-      <origin rpy="0 0 0" xyz="0 0 0.15"/>
-      <parent link="link7"/>
-      <child link="end-effector"/>
-    </joint>
-</robot>
