--- examples/21-singularity/world.urdf
@@ -1,38 +1,0 @@
-<?xml version="1.0" ?>
-
-<world name="demo_world" gravity="0.0 0.0 0.0">
-
-	<robot name="PANDA">
-		<model dir="./resources" path="panda_arm.urdf" name="panda" />
-	</robot>
-
-	<static_object name="Floor">
-		<origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-	    <visual>
-	        <origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-	        <geometry>
-	            <box size="3.0 3.0 0.1" />
-	        </geometry>
-	        <material name="material_red">
-		        <color rgba="0.1 0.1 0.1 1.0" />
-	        </material>
-	    </visual>
-	</static_object>
-
-	<light name="light1" type="directional">
-		<position xyz="2.0 -2.0 2.0" />
-		<lookat xyz="0.0 0.0 0.0" />
-	</light>
-
-	<light name="light2" type="directional">
-		<position xyz="2.0 2.0 2.0" />
-		<lookat xyz="0.0 0.0 0.0" />
-	</light>
-
-	<camera name="camera">
-		<position xyz="1.5 0.4 1.0" />
-		<vertical xyz="0.0 0.0 1.0" />
-		<lookat xyz="0.0 0.0 0.5" />
-	</camera>
-
-</world>
