--- examples/22-puma_singularity/world.urdf
@@ -1,48 +1,0 @@
-<?xml version="1.0" ?>
-
-<world name="demo_world" gravity="0.0 0.0 0.0">
-<!-- <world name="demo_world" gravity="0.0 0.0 -9.81"> -->
-
-	<robot name="PUMA">
-		<!-- model node is required -->
-		<model dir="./resources" path="puma.urdf" name="Puma" />
-		<!-- origin node is optional -->
-		<origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-	</robot>
-
-	<static_object name="Floor">
-		<origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-	    <visual>
-	        <origin xyz="0.0 0.0 0.0" rpy="0 0 0" />
-	        <geometry>
-	            <box size="3.0 3.0 0.1" />
-	        </geometry>
-	        <material name="material_red">
-		        <color rgba="0.1 0.1 0.1 1.0" />
-	        </material>
-	    </visual>
-	</static_object>
-
-	<light name="light1" type="directional">
-		<position xyz="-2.0 -2.0 2.0" />
-		<lookat xyz="0.0 0.0 0.0" />
-	</light>
-
-	<light name="light2" type="directional">
-		<position xyz="-2.0 2.0 2.0" />
-		<lookat xyz="0.0 0.0 0.0" />
-	</light>
-
-	<!-- <camera name="camera">
-		<position xyz="-2 0.5 0.75" />
-		<vertical xyz="0.0 0.0 1.0" />
-		<lookat xyz="0.0 0.5 0.75" />
-	</camera> -->
-
-	<camera name="camera">
-		<position xyz="-1.34848 0.766375 1.18243" />
-		<vertical xyz="0.401548 -0.133621  0.906038" />
-		<lookat xyz="-0.469794  0.543684  0.760161" />
-	</camera>
-
-</world>
