+++ examples/CMakeLists.txt
@@ -45,5 +45,5 @@ add_subdirectory(11-planar_robot_controller)
 add_subdirectory(15-haptic_control_impedance_type)
 add_subdirectory(16-haptic_control_admittance_type)
 add_subdirectory(17-bilateral_teleop_with_POPC)
-add_subdirectory(21-singularity)
-add_subdirectory(22-puma_singularity)
+add_subdirectory(18-singularity)
+add_subdirectory(19-puma_singularity)
