+++ src/tasks/SingularityHandler.cpp
@@ -31,9 +31,10 @@ SingularityHandler::SingularityHandler(std::shared_ptr<Sai2Model::Sai2Model> rob
                                        _type_1_tol(type_1_tol), 
                                        _type_2_torque_ratio(type_2_torque_ratio), 
                                        _perturb_step_size(perturb_step_size),
+                                       _buffer_size(buffer_size),
                                        _verbose(verbose)
 {
-    // initialize robot specific variables 
+    // initialize limits 
     _q_upper = VectorXd::Zero(_robot->dof());
     _q_lower = VectorXd::Zero(_robot->dof());
     _tau_upper = VectorXd::Zero(_robot->dof());
@@ -57,17 +58,14 @@ SingularityHandler::SingularityHandler(std::shared_ptr<Sai2Model::Sai2Model> rob
     setGains(50, 14, 5);
     _type_1_counter = 0;
     _type_2_counter = 0;
-    _singular_task_force = VectorXd::Zero(6);
-    _singular_task_torques = VectorXd::Zero(_robot->dof());
-    _joint_strategy_torques = VectorXd::Zero(_robot->dof());
     _type_2_direction = VectorXd::Ones(_robot->dof());
+    _enforce_type_1_strategy = false;
 }
 
 void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, const MatrixXd& N_prec) {
     
     // task range decomposition
     int dof = _robot->dof();
-
     JacobiSVD<MatrixXd> J_svd(projected_jacobian, ComputeThinU | ComputeThinV);
     _svd_U = J_svd.matrixU();
     _svd_s = J_svd.singularValues();
@@ -77,7 +75,7 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
         // fully singular task
         _alpha = 0;
 
-        // non-singular task
+        // placeholder non-singular task range 
         _task_range_ns = MatrixXd::Zero(_task_rank, 1);
 
         // singular task 
@@ -89,11 +87,11 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
 					_projected_jacobian_s.transpose()).completeOrthogonalDecomposition().pseudoInverse();
     } else {
         for (int i = 1; i < _task_rank; ++i) {
-            double condition_number = _svd_s(i) / _svd_s(0);
+            double inv_condition_number = _svd_s(i) / _svd_s(0);
 
-            if (condition_number < _s_max) {
+            if (inv_condition_number < _s_max) {
                 // task enters singularity blending region
-                _alpha = std::clamp((condition_number - _s_min) / (_s_max - _s_min), 0., 1.);
+                _alpha = std::clamp((inv_condition_number - _s_min) / (_s_max - _s_min), 0., 1.);
 
                 // non-singular task
                 _task_range_ns = _svd_U.leftCols(i);
@@ -104,14 +102,11 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
                 _Jbar_ns = ns_matrices.Jbar;
                 _N_ns = ns_matrices.N;
 
-                // singular task 
-                // task range only collects columns of U up to size task_rank - non-singular task rank
+                // singular task: task range only collects columns of U up to size task_rank - non-singular task rank
                 _task_range_s = _svd_U.block(0, i, _svd_U.rows(), _task_rank - i);  
                 _joint_task_range_s = _svd_V.block(0, i, _svd_V.rows(), _task_rank - i);
-                _projected_jacobian_s = _task_range_s.transpose() * projected_jacobian;  // in N_ns?
-                Sai2Model::OpSpaceMatrices s_matrices =
-                    _robot->operationalSpaceMatrices(_projected_jacobian_s);
-                _Lambda_s = s_matrices.Lambda;
+                _projected_jacobian_s = _task_range_s.transpose() * projected_jacobian;  
+                _Lambda_s = (_projected_jacobian_s * _robot->MInv() * _projected_jacobian_s.transpose()).inverse();
                 break;
 
             } else if (i == _task_rank - 1) {
@@ -127,26 +122,24 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
                 _Jbar_ns = ns_matrices.Jbar;
                 _N_ns = ns_matrices.N;
 
-                // singular task 
+                // placeholder singular task range
                 _task_range_s = MatrixXd::Zero(_task_rank, 1);
                 _joint_task_range_s = MatrixXd::Zero(dof, 1);
             }
         }
     }
 
-    // task update
+    // model updates 
     if (_task_range_s.norm() == 0) {
-        _N = _N_ns;  // _N = N_ns (rank(N_ns) = task_rank)
+        _N = _N_ns;  
     } else {
-        // update posture task (occupies all singular joint directions, but joint torque handles one at a time)
         _posture_projected_jacobian = _joint_task_range_s.transpose() * _N_ns * N_prec;
         Sai2Model::OpSpaceMatrices op_space_matrices =
             _robot->operationalSpaceMatrices(_posture_projected_jacobian);
         _Lambda_joint_s = op_space_matrices.Lambda;
-        _N = op_space_matrices.N * _N_ns;  // _N = N_partial_joint * N_Ns (rank(_N = N_partial_joint * N_ns) = task_rank)
+        _N = op_space_matrices.N * _N_ns; 
     }
 
-    // apply dynamic decoupling modifications
     switch (_dynamic_decoupling_type) {
 		case FULL_DYNAMIC_DECOUPLING: {
 			_Lambda_ns_modified = _Lambda_ns;
@@ -171,7 +164,7 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
 			}
 			MatrixXd M_inv_BIE = M_BIE.inverse();
 
-			// nonsingular lambda
+			// non-singular lambda
 			MatrixXd Lambda_inv_BIE =
 				_projected_jacobian_ns *
 				M_inv_BIE * 
@@ -192,8 +185,8 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
             // joint strategy lambda 
             if (_task_range_s.norm() != 0) {
                 Lambda_inv_BIE = _posture_projected_jacobian * 
-                                    M_inv_BIE * 
-                                    _posture_projected_jacobian.transpose();
+                                 M_inv_BIE * 
+                                 _posture_projected_jacobian.transpose();
                 _Lambda_joint_s_modified = Lambda_inv_BIE.inverse();
             } 
 			break;
@@ -207,7 +200,6 @@ void SingularityHandler::updateTaskModel(const MatrixXd& projected_jacobian, con
 		}
 	}
 
-    // classify singularity
     classifySingularity(_task_range_s, _joint_task_range_s);
 
 }
@@ -226,7 +218,6 @@ void SingularityHandler::classifySingularity(const MatrixXd& singular_task_range
         _singularity_history.clear();
         _type_1_counter = 0;
         _type_2_counter = 0;
-        _singularity_task_range_history.clear();
         return;
     }
 
@@ -258,16 +249,13 @@ void SingularityHandler::classifySingularity(const MatrixXd& singular_task_range
         _robot->updateKinematics();
     }
 
-    // add to buffer and counters (preference for handling type 1 over type 2 for multiple singularities)
+    // add to buffer and counters (preference for handling type 1 over type 2 for multiple, simultaneous singularities)
     auto it = std::find(_singularity_types.begin(), _singularity_types.end(), TYPE_1_SINGULARITY);
     if (it != _singularity_types.end()) {
-        int index = std::distance(_singularity_types.begin(), it);
         _singularity_history.push_back(TYPE_1_SINGULARITY);
-        _singularity_task_range_history.push_back(singular_task_range.col(index));
         _type_1_counter++;
     } else {
         _singularity_history.push_back(TYPE_2_SINGULARITY);
-        _singularity_task_range_history.push_back(singular_task_range.col(0));
         _type_2_counter++;
     }
 
@@ -279,13 +267,11 @@ void SingularityHandler::classifySingularity(const MatrixXd& singular_task_range
             _type_2_counter--;
         }
         _singularity_history.pop_front();
-        _singularity_task_range_history.pop_front();
     }
 
 }
 
-VectorXd SingularityHandler::computeTorques(const VectorXd& unit_mass_force, const VectorXd& force_related_terms)
-{
+VectorXd SingularityHandler::computeTorques(const VectorXd& unit_mass_force, const VectorXd& force_related_terms) {
     if (_verbose) {
         if (_singularity_types.size() != 0) {
             for (auto type : _singularity_types) {
@@ -296,75 +282,57 @@ VectorXd SingularityHandler::computeTorques(const VectorXd& unit_mass_force, con
     }
 
     if (_singularity_types.size() == 0) {
-        VectorXd tau = _projected_jacobian_ns.transpose() * (_Lambda_ns_modified * _task_range_ns.transpose() * unit_mass_force + \
-                            _task_range_ns.transpose() * force_related_terms);
-        // logging 
-        _nonsingular_task_force = _Lambda_ns_modified * _task_range_ns.transpose() * unit_mass_force;
-        _singular_task_force = VectorXd::Zero(1);
-        _singular_task_torques = VectorXd::Zero(_robot->dof());
-        _joint_strategy_torques = VectorXd::Zero(_robot->dof());
-        _original_torques = tau;
-        _blended_torques = tau;
-        _singular_torques = VectorXd::Zero(7);
-        _nonsingular_torques = tau;
-        return tau;
+        return _projected_jacobian_ns.transpose() * (_Lambda_ns_modified * _task_range_ns.transpose() * unit_mass_force + \
+                    _task_range_ns.transpose() * force_related_terms);
     } else {
-        VectorXd unit_torques = VectorXd::Zero(_robot->dof());
         VectorXd tau_ns = VectorXd::Zero(_robot->dof());
 
         // compute non-singular torques 
         if (_task_range_ns.norm() != 0) {
             tau_ns = _projected_jacobian_ns.transpose() * (_Lambda_ns_modified * _task_range_ns.transpose() * unit_mass_force + \
                         _task_range_ns.transpose() * force_related_terms);
-            _nonsingular_task_force = _Lambda_ns_modified * _task_range_ns.transpose() * unit_mass_force;
         }
 
-        // handle singularity type based on which one has more coutns 
-        if (_type_1_counter > _type_2_counter) {
+        // handle singularity type based on which one has more counts  
+        if (_type_1_counter > _type_2_counter || _enforce_type_1_strategy) {
             // joint holding to entering joint conditions  
-            unit_torques = - _kp * (_robot->q() - _q_prior) - _kv * _robot->dq();  
+            VectorXd unit_torques = - _kp_type_1 * (_robot->q() - _q_prior) - _kv_type_1 * _robot->dq();  
             _joint_strategy_torques = _posture_projected_jacobian.transpose() * _Lambda_joint_s_modified * _joint_task_range_s.transpose() * unit_torques;
         } else {
             // apply open-loop torque proportional to dot(unit mass force, singular direction)
-            // zero torque achieved when singular direction is orthgonal to the desired unit mass force direction
-            // the direction is chosen away from joint limit if joint is in limit 
-            _fTd = ((unit_mass_force.normalized() + force_related_terms.normalized()).normalized()).dot(_task_range_s.col(0));
-            VectorXd magnitude_unit_torques = std::abs(_fTd) * _type_2_torque_vector;
+            // zero torque achieved when singular direction is orthogonal to the desired unit mass force direction
+            // the direction is reversed if the joint is approaching a joint limit 
             for (int i = 0; i < _joint_task_range_s.rows(); ++i) {
                 if (_joint_task_range_s(i, 0) != 0) {
                     if (std::abs(_robot->q()(i) - _q_upper(i)) < 5 * M_PI / 180) {
-                        _type_2_direction(i) = -1;
+                        _type_2_direction(i) = - 1;
                     } else if (std::abs(_robot->q()(i) - _q_lower(i)) < 5 * M_PI / 180) {
                         _type_2_direction(i) = 1;
                     } 
                 }
             }
-            unit_torques = _type_2_direction.array() * magnitude_unit_torques.array(); 
+            double fTd = ((unit_mass_force.normalized() + force_related_terms.normalized()).normalized()).dot(_task_range_s.col(0));
+            VectorXd magnitude_unit_torques = std::abs(fTd) * _type_2_torque_vector;
+            VectorXd unit_torques = _type_2_direction.array() * magnitude_unit_torques.array(); 
             _joint_strategy_torques = _posture_projected_jacobian.transpose() * _joint_task_range_s.transpose() * unit_torques + \
-                                        _posture_projected_jacobian.transpose() * _Lambda_joint_s_modified * _joint_task_range_s.transpose() * (- _kv_type_2 * _robot->dq());
+                                        _posture_projected_jacobian.transpose() * _Lambda_joint_s_modified * \
+                                        _joint_task_range_s.transpose() * (- _kv_type_2 * _robot->dq());
         }
         
-        // combine non-singular torques and blended singular torques with joint strategy torques (add joint space damping in singular task force)
-        _singular_task_force = _Lambda_s_modified * _task_range_s.transpose() * unit_mass_force + \
+        // combine non-singular torques and blended singular torques with joint strategy torques
+        _singular_task_torques = _projected_jacobian_s.transpose() * _Lambda_s_modified * _task_range_s.transpose() * unit_mass_force + \
                                             _task_range_s.transpose() * force_related_terms;
-        _singular_task_torques = _projected_jacobian_s.transpose() * _singular_task_force;
 
         for (int i = 0; i < _robot->dof(); ++i) {
             if (isnan(_singular_task_torques(i))) {
-                _singular_task_torques(i) = _singular_task_torques(i) > 0 ? _tau_upper(i) : _tau_lower(i);
+                _singular_task_torques(i) = 0;  
             } else if (_singular_task_torques(i) > _tau_upper(i)) {
                 _singular_task_torques(i) = _tau_upper(i);
             } else if (_singular_task_torques(i) < _tau_lower(i)) {
                 _singular_task_torques(i) = _tau_lower(i);
             }
         }
-        VectorXd tau_s = pow(_alpha, 1) * _singular_task_torques + (1 - pow(_alpha, 1)) * _joint_strategy_torques;
-        _original_torques = tau_ns + _singular_task_torques;
-        _blended_torques = tau_ns + tau_s;
-        _nonsingular_torques = tau_ns;
-        _singular_torques = tau_s;
-
-        return tau_ns + tau_s;
+        return tau_ns + _alpha * _singular_task_torques + (1 - _alpha) * _joint_strategy_torques;
     }
 }
