+++ src/tasks/SingularityHandler.h
@@ -27,64 +27,11 @@ enum SingularityType {
     TYPE_1_SINGULARITY,  
     TYPE_2_SINGULARITY
 };
-struct SingularityHandlerData {
-    std::vector<SingularityType> singularity_types;
-    double alpha;
-    VectorXd task_range_s;
-    VectorXd s;
-    VectorXd joint_task_range_s;
-    VectorXd singular_task_force;
-    VectorXd nonsingular_task_force;
-    VectorXd singular_task_torques;
-    VectorXd joint_strategy_torques;
-    VectorXd original_torques;
-    VectorXd nonsingular_torques;
-    VectorXd singular_torques;
-    VectorXd blended_torques;
-    double fTd;
-    SingularityHandlerData(const std::vector<SingularityType> singularity_types,
-                           const double& alpha,
-                           const VectorXd& task_range_s,
-                           const VectorXd& s,
-                           const VectorXd& joint_task_range_s,
-                           const VectorXd& singular_task_force,
-                           const VectorXd& nonsingular_task_force,
-                           const VectorXd& singular_task_torques,
-                           const VectorXd& joint_strategy_torques,
-                           const VectorXd& original_torques,
-                           const VectorXd& nonsingular_torques,
-                           const VectorXd& singular_torques,
-                           const VectorXd& blended_torques,
-                           const double& fTd) : 
-                           singularity_types(singularity_types),
-                           alpha(alpha),
-                           task_range_s(task_range_s),
-                           s(s),
-                           joint_task_range_s(joint_task_range_s),
-                           singular_task_force(singular_task_force),
-                           nonsingular_task_force(nonsingular_task_force),
-                           singular_task_torques(singular_task_torques),
-                           joint_strategy_torques(joint_strategy_torques),
-                           original_torques(original_torques),
-                           nonsingular_torques(nonsingular_torques),
-                           singular_torques(singular_torques),
-                           blended_torques(blended_torques),
-                           fTd(fTd) {};
-};
-
-struct SvdData {
-    MatrixXd U, V;
-    VectorXd s;
-    SvdData(const MatrixXd& U,
-            const VectorXd& s,
-            const MatrixXd& V) : 
-            U(U), s(s), V(V) {}
-};
 
 const std::vector<std::string> singularity_labels {"No Singularity", "Type 1 Singularity", "Type 2 Singularity"};      
 
 enum DynamicDecouplingType {
-    FULL_DYNAMIC_DECOUPLING = 0,	 // use the real Lambda matrix
+    FULL_DYNAMIC_DECOUPLING = 0, // use the real Lambda matrix
     IMPEDANCE,					 // use Identity for the mass matrix
     BOUNDED_INERTIA_ESTIMATES,	 // use a Lambda computed from a saturated
                                  // joint space mass matrix
@@ -101,8 +48,7 @@ public:
      * @param task_rank rank of the motion force task after partial task projection
      * @param s_abs_tol tolerance to declare task completely singular if all singular values from the projected
      * jacobian are under this value
-     * @param type_1_tol tolerance for the dot product between the first and last singular direction in the 
-     * singular direction buffer for which a type 1 singularity is classified
+     * @param type_1_tol tolerance to classify type 1 singularities 
      * @param type_2_torque_ratio percentage of the max joint torque to use for the type 2 singularity strategy
      * @param perturb_step_size the joint angle increment to classify singularity 
      * @param buffer_size size of buffer to compare which singularity type has more counts 
@@ -117,7 +63,7 @@ public:
                        const double& type_2_torque_ratio = 1e-2,
                        const double& perturb_step_size = 5,
                        const int& buffer_size = 200,
-                       const bool& verbose = true);
+                       const bool& verbose = false);
 
     /**
      * @brief Updates the model quantities for the singularity handling task, and performs singularity classification
@@ -128,10 +74,7 @@ public:
     void updateTaskModel(const MatrixXd& projected_jacobian, const MatrixXd& N_prec);
 
     /**
-     * @brief Classifies the singularity based on a joint perturbation in the J_{s} N_{ns} direction since 
-     * \delta x = J_{ns} \delta q + J_{s} N_{ns} \delta q_{0}. Type 1 singularity if the norm of 
-     * \delta x is greater than the type 1 tolerance, otherwise type 2. Type 2 singularity will have a very
-     * small \delta x norm.
+     * @brief Classifies the singularity based on a joint perturbation in the singular joint space 
      * 
      * @param singular_task_range Singular task range corresponding to the columns of U from SVD
      * @param singular_joint_task_range Singular task range corresponding to the columns of V from SVD
@@ -140,9 +83,9 @@ public:
                              const MatrixXd& singular_joint_task_range);
 
     /**
-     * @brief Computes the torques from the singularity handling. If the projected jacobian doesn't have
-     * a condition number below the _s_max value, then the torque is computed as usual with _projected_jacobian_ns.
-     * If the projected jacobian has a condition number below the _s_max value, then the torque is computed as
+     * @brief Computes the torques from the singularity handling. If the projected jacobian isn't classified singular, then
+     * the torque is computed as usual.
+     * If the projected jacobian is classified singular, then the torque is computed as
      * \tau = \tau_{ns} + (1 - \_alpha) * \tau_{joint strategy} + \alpha * \tau_{s} where \alpha is the linear blending ratio, 
      * \tau_{ns} is the torque computed from the non-singular terms, \tau_{s} is the torque computed from the singular 
      * terms, and \tau_{joint strategy} is the torque computed from the singularity strategy.
@@ -163,37 +106,19 @@ public:
     }
 
     /**
-     * @brief Get the nullspace of the singularity task
+     * @brief Get the nullspace 
      * 
      * @return MatrixXd nullspace 
      */
     MatrixXd getNullspace() { return _N; };
 
     /**
-     * @brief Get the Svd Data object
-     * 
-     * @return MatrixXd 
-     */
-    SvdData getSvdData() {
-        return SvdData(_svd_U, _svd_s, _svd_V);
-    }
-
-    /**
-     * @brief Get the non-singular op-space matrices
-     * 
-     * @return Sai2Model::OpSpaceMatrices non-singular op-space matrices
-     */
-    Sai2Model::OpSpaceMatrices getNonSingularOpSpaceMatrices() {
-        return Sai2Model::OpSpaceMatrices(_projected_jacobian_ns, _Lambda_ns, _Jbar_ns, _N_ns);
-    }
-
-    /**
-     * @brief Set the singularity bounds for torque blending based on the condition number
+     * @brief Set the singularity bounds for torque blending based on the inverse of the condition number
      * The linear blending coefficient \alpha is computed as \alpha = (s - _s_min) / (_s_max - _s_min),
      * and is clamped between 0 and 1.
      * 
-     * @param s_min condition number value to only use singularity strategy torques
-     * @param s_max condition number value to start blending
+     * @param s_min lower bound
+     * @param s_max upper bound 
      */
     void setSingularityBounds(const double& s_min, const double& s_max) {
         _s_min = s_min;
@@ -203,54 +128,24 @@ public:
     /**
      * @brief Set the gains for the partial joint task for the singularity strategy
      * 
-     * @param kp position gain
-     * @param kv velocity gain
+     * @param kp_type_1 position gain for type 1 strategy
+     * @param kv_type_1 velocity damping gain for type 1 strategy
+     * @param kv_type_2 velocity damping gain for type 2 strategy
      */
-    void setGains(const double& kp, const double& kv, const double& kv_type_2) {
-        _kp = kp;
-        _kv = kv;
+    void setGains(const double& kp_type_1, const double& kv_type_1, const double& kv_type_2) {
+        _kp_type_1 = kp_type_1;
+        _kv_type_1 = kv_type_1;
         _kv_type_2 = kv_type_2;
     }
 
     /**
-     * @brief Set the torque ratio used for type 2 singularity strategy
+     * @brief Enforces type 1 handling behavior if set to true, otherwise handle 
+     * type 1 or type 2 as usual
      * 
-     * @param ratio Ratio of the specific joint's max torque to use 
+     * @param flag  true to enforce type 1 handling behavior 
      */
-    void setClassificationParams(const double& perturb_step_size, 
-                                 const double& type_1_tol,
-                                 const double& type_2_torque_ratio = 1e-2) {
-        _perturb_step_size = perturb_step_size;
-        _type_1_tol = type_1_tol;
-        _type_2_torque_ratio = type_2_torque_ratio;
-    }
-
-    /**
-     * @brief Return singularity handler data 
-    */
-    SingularityHandlerData getData() {
-        return SingularityHandlerData(_singularity_types,
-                                      _alpha,
-                                      _task_range_s,
-                                      _svd_s,
-                                      _joint_task_range_s,
-                                      _singular_task_force,
-                                      _nonsingular_task_force,
-                                      _singular_task_torques,
-                                      _joint_strategy_torques,
-                                      _original_torques,
-                                      _nonsingular_torques,
-                                      _singular_torques,
-                                      _blended_torques,
-                                      _fTd);
-    }
-
-    MatrixXd getNonSingularLambda() {
-        return _Lambda_ns_modified;
-    }
-
-    MatrixXd getSingularLambda() {
-        return _Lambda_s_modified;
+    void setType1Behavior(const bool flag) {
+        _enforce_type_1_strategy = true;
     }
 
 private:
@@ -261,36 +156,36 @@ private:
     Affine3d _compliant_frame;
     int _task_rank;
     VectorXd _joint_midrange, _q_upper, _q_lower, _tau_upper, _tau_lower;
+    bool _enforce_type_1_strategy;
     bool _verbose;
 
     // singularity information
     std::vector<SingularityType> _singularity_types;
     double _perturb_step_size;
     std::deque<SingularityType> _singularity_history;
-    std::deque<MatrixXd> _singularity_task_range_history;
     int _type_1_counter, _type_2_counter;
     int _buffer_size;
 
     // type 1 specifications
     VectorXd _q_prior, _dq_prior;
-    double _kp, _kv;
+    double _kp_type_1, _kv_type_1;
     double _type_1_tol;
 
     // type 2 specifications
     double _type_2_torque_ratio;  // use X% of the max joint torque 
-    VectorXd _type_2_torque_vector;
     double _kv_type_2;
+    VectorXd _type_2_torque_vector;
     VectorXd _type_2_direction;
-    double _fTd;
 
     // model quantities 
+    MatrixXd _svd_U, _svd_V;
+    VectorXd _svd_s;
     double _s_abs_tol;  
     double _s_min, _s_max;
     double _alpha;
     MatrixXd _N;
     MatrixXd _task_range_ns, _task_range_s, _joint_task_range_s;
     MatrixXd _projected_jacobian_ns, _projected_jacobian_s;
-    MatrixXd _non_truncated_projected_jacobian_s;
     MatrixXd _Lambda_ns, _Jbar_ns, _N_ns;
     MatrixXd _Lambda_s;
     MatrixXd _Lambda_ns_modified, _Lambda_s_modified;
@@ -298,18 +193,9 @@ private:
 
     // joint task quantities 
     MatrixXd _posture_projected_jacobian, _M_partial;
-
-    // singularity handler   
-    MatrixXd _svd_U, _svd_V;
-    VectorXd _svd_s;
-    VectorXd _singular_task_force;
-    VectorXd _nonsingular_task_force;
+    
     VectorXd _singular_task_torques;
     VectorXd _joint_strategy_torques;
-    VectorXd _original_torques;
-    VectorXd _blended_torques;
-    VectorXd _nonsingular_torques;
-    VectorXd _singular_torques;
 };
 
 }  // namespace
